requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.3
postgres==4.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
